requirements.txt
streamlit==1.32.2

langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.36
langchain-openai==0.1.7
langchain-chroma==0.1.0
langchain-text-splitters==0.0.1

langchain-groq==0.1.3
langchain-huggingface==0.0.3

chromadb==0.4.24
faiss-cpu==1.7.4

sentence-transformers==2.7.0
pypdf==4.2.0
python-dotenv==1.0.1
tiktoken==0.7.0